+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog of Cura-DuetRRFPlugin
 
+## v1.2.3: 2021-01-30
+  * move deleting of unmapped settings to action button on message
+  * correctly bump plugin version metadata
+
 ## v1.2.2: 2021-01-27
   * add OutputDevices on currently active printer after saving config
   * remove OutputDevices on currently active printer after deleting config